[BUGFIX] Replaced Long method with Integer

+++ src/main/java/org/rev317/min/api/methods/Skill.java
@@ -32,7 +32,7 @@ public enum Skill {
     SUMMONING,
     DUNGEONEERING;
 
-    private static final long[] EXPERIENCE = {
+    private static final int[] EXPERIENCE = {
             0, 0, 83, 174, 276, 388, 512, 650, 801, 969, 1154, 1358, 1584,
             1833, 2107, 2411, 2746, 3115, 3523, 3973, 4470, 5018, 5624, 6291,
             7028, 7842, 8740, 9730, 10824, 12031, 13363, 14833, 16456, 18247,
@@ -51,13 +51,12 @@ public enum Skill {
             280681209, 309897078, 342154009, 377768545, 417090179, 460504778,
             508438379, 561361362, 619793069, 684306901, 755535943, 834179178,
             921008346, 1016875516, 1122721449, 1239584831, 1368612462, 1511070513,
-            1668356950, 1842015252, 2033749558 };
+            1668356950, 1842015252, 2033749558};
 
     /**
      * Returns the experience of the provided skill.
      *
      * @param index the skill index.
-     *
      * @return the experience.
      */
     public static final int getCurrentExperience(int index) {
@@ -68,7 +67,6 @@ public enum Skill {
      * Returns the real level of the provided skill.
      *
      * @param index the skill index.
-     *
      * @return the real skill level.
      */
     public static final int getRealLevel(int index) {
@@ -79,7 +77,6 @@ public enum Skill {
      * Returns the current level of the provided skill. (Will return de-buffed/buffed levels)
      *
      * @param index the skill index.
-     *
      * @return the current skill level. Done by Bears
      */
     public static final int getCurrentLevel(int index) {
@@ -90,10 +87,9 @@ public enum Skill {
      * Returns the exact experience at the provided level.
      *
      * @param level the level.
-     *
      * @return the experience at the provided level.
      */
-    public static final long getExperienceByLevel(int level) {
+    public static final int getExperienceByLevel(int level) {
         if (level > 150 || level < 1) {
             return 0;
         }
@@ -105,7 +101,6 @@ public enum Skill {
      * Returns the exact level with the provided experience.
      *
      * @param experience the experience.
-     *
      * @return the level at the provided experience.
      */
     public static final int getLevelByExperience(int experience) {
@@ -122,10 +117,9 @@ public enum Skill {
      * Returns the remaining experience for the provided skill to level up.
      *
      * @param index the skill index.
-     *
      * @return the remaining experience.
      */
-    public static final long getRemainingExperience(int index) {
+    public static final int getRemainingExperience(int index) {
         int level = getLevelByExperience(getCurrentExperience(index));
         if (level > 150 || level < 1) {
             return 0;
@@ -138,12 +132,11 @@ public enum Skill {
      * Returns the percentage to the next level for the provided skill.
      *
      * @param index the skill index.
-     *
      * @return the remaining percentage.
      */
     public static final int getPercentToNextLevel(int index) {
         int currentLevel = getLevelByExperience(getCurrentExperience(index));
-        int nextLevel    = currentLevel + 1;
+        int nextLevel = currentLevel + 1;
         if (currentLevel == 150 || nextLevel > 150 || currentLevel < 1
                 || nextLevel < 1) {
             return 0;
@@ -191,7 +184,7 @@ public enum Skill {
     /**
      * Returns the remaining experience until the next level.
      */
-    public final long getRemaining() {
+    public final int getRemaining() {
         return Skill.getRemainingExperience(this.ordinal());
     }
